requirements.txt
lxml==4.5.0
python-dateutil==2.8.1
pytz==2019.3
boto3==1.12.47
bigquery-schema-generator==1.0
apache-airflow[crypto,celery,postgres,hive,jdbc,ssh]==1.10.10
dask[complete]==2.15.0
distributed==2.15.0
google-cloud-bigquery==1.24.0
six==1.14.0
